+++ example/lib/my_icons.dart
@@ -1,0 +1,40 @@
+// Generated code: do not hand-edit.
+
+// Generated using fontify_plus.
+// Copyright © 2026 fontify_plus (https://github.com/4akloon/fontify_plus).
+
+import 'package:flutter/widgets.dart';
+
+/// Identifiers for the icons.
+///
+/// Use with the [Icon] class to show specific icons.
+///
+/// Icons are identified by their name as listed below.
+///
+/// To use this class, make sure you declare the font in your
+/// project's `pubspec.yaml` file in the `fonts` section. This ensures that
+/// the "My Icons" font is included in your application. This font is used
+/// to display the icons. For example:
+///
+/// ```yaml
+/// flutter:
+///   fonts:
+///     - family: My Icons
+///       fonts:
+///         - asset: fonts/my_icons.otf
+/// ```
+abstract final class MyIcons {
+  static const iconFontFamily = 'My Icons';
+
+  /// arrow_right
+  static const IconData arrowRight = IconData(0xe000, fontFamily: iconFontFamily);
+
+  /// plus
+  static const IconData plus = IconData(0xe001, fontFamily: iconFontFamily);
+
+  /// check
+  static const IconData check = IconData(0xe002, fontFamily: iconFontFamily);
+
+  /// menu
+  static const IconData menu = IconData(0xe003, fontFamily: iconFontFamily);
+}